Organizes thoughts into clear architecture
